social-crawler/requirements.txt
httpx>=0.25.0
aiohttp>=3.9.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
feedparser>=6.0.0
nltk>=3.8.0
